fix: minimizing the dependencies for the rest module
closes: #48114
