+++ rest/admin-v2/api/src/main/java/org/keycloak/rest/admin/api/DefaultAdminRootV2.java
@@ -1,10 +1,12 @@
-package org.keycloak.admin.api;
+package org.keycloak.rest.admin.api;
 
 import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.ext.Provider;
 
+import org.keycloak.admin.api.AdminApi;
+import org.keycloak.admin.api.AdminRootV2;
 import org.keycloak.common.Profile;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.services.resources.admin.AdminCorsPreflightService;
