+++ rest/admin-v2/api/src/main/resources/META-INF/services/jakarta.validation.ConstraintValidator
@@ -1,0 +1,3 @@
+org.keycloak.representations.admin.v2.validation.impl.ClientSecretNotBlankValidator
+org.keycloak.representations.admin.v2.validation.impl.ProtocolUnmodifiedValidator
+org.keycloak.representations.admin.v2.validation.impl.UuidUnmodifiedValidator
