+++ rest/admin-v2/rest/src/main/java/org/keycloak/admin/api/PatchTypeNames.java
@@ -1,4 +1,4 @@
-package org.keycloak.services;
+package org.keycloak.admin.api;
 
 public class PatchTypeNames {
     public static final String JSON_MERGE = "application/merge-patch+json";
