+++ rest/admin-v2/rest/src/main/java/org/keycloak/representations/admin/v2/validation/ProtocolUnmodified.java
@@ -11,8 +11,8 @@ import jakarta.validation.Payload;
 
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = {ProtocolUnmodifiedValidator.class})
 @Documented
+@Constraint(validatedBy = {})
 public @interface ProtocolUnmodified {
     String message() default "protocol cannot be changed for an existing client";
     Class<?>[] groups() default {};
