+++ rest/admin-v2/rest/src/main/java/org/keycloak/representations/admin/v2/validation/PutClient.java
