+++ services/src/main/java/org/keycloak/services/resources/KeycloakOpenAPI.java
@@ -17,49 +17,8 @@
 
 package org.keycloak.services.resources;
 
-/**
- * Class of constants relating to the OpenAPI annotations in Keycloak and the Keycloak Admin REST API
- */
-public class KeycloakOpenAPI {
+public class KeycloakOpenAPI extends org.keycloak.common.constants.KeycloakOpenAPI {
 
     private KeycloakOpenAPI() { }
-    public static class Profiles {
 
-        public static final String ADMIN = "x-smallrye-profile-admin";
-
-        private Profiles() { }
-    }
-
-    public static class Admin {
-
-        private Admin() { }
-
-        public static class Tags {
-            public static final String ATTACK_DETECTION = "Attack Detection";
-            public static final String AUTHENTICATION_MANAGEMENT = "Authentication Management";
-            public static final String CLIENTS = "Clients";
-            public static final String CLIENTS_V2 = "Clients (v2)";
-            public static final String CLIENT_ATTRIBUTE_CERTIFICATE = "Client Attribute Certificate";
-            public static final String CLIENT_INITIAL_ACCESS = "Client Initial Access";
-            public static final String CLIENT_REGISTRATION_POLICY = "Client Registration Policy";
-            public static final String CLIENT_ROLE_MAPPINGS = "Client Role Mappings";
-            public static final String CLIENT_SCOPES = "Client Scopes";
-            public static final String COMPONENT = "Component";
-            public static final String GROUPS = "Groups";
-            public static final String IDENTITY_PROVIDERS = "Identity Providers";
-            public static final String KEY = "Key";
-            public static final String PROTOCOL_MAPPERS = "Protocol Mappers";
-            public static final String REALMS_ADMIN = "Realms Admin";
-            public static final String ROLES = "Roles";
-            public static final String ROLES_BY_ID = "Roles (by ID)";
-            public static final String ROLE_MAPPER = "Role Mapper";
-            public static final String ROOT = "Root";
-            public static final String SCOPE_MAPPINGS = "Scope Mappings";
-            public static final String USERS = "Users";
-            public static final String ORGANIZATIONS = "Organizations";
-            public static final String WORKFLOWS = "Workflows";
-            private Tags() { }
-        }
-
-    }
 }
